+++ src/bugit_v2/screens/submission_progress_screen.py
@@ -410,8 +410,12 @@ class SubmissionProgressScreen[TAuth](Screen[ReturnScreenChoice]):
         if not self.finished:
             return
 
+        # immediately hide the give up button
+        self.query_exactly_one("#give_up", Button).display = False
+
         try:
             rv = self.submitter.finalize()
+            finalize_ok = True
             if rv:
                 self._log_with_time(f"[green]FINALIZE OK[/] {rv}")
             else:
@@ -419,25 +423,27 @@ class SubmissionProgressScreen[TAuth](Screen[ReturnScreenChoice]):
                     f"[green]FINALIZE OK[/] {self.submitter.display_name}"
                 )
         except Exception as e:
+            finalize_ok = False
             self._log_with_time(f"[red]ERR when finalizing[/]: {repr(e)}")
             logger.error(e)
 
-        self.query_exactly_one("#give_up", Button).display = False
+        finish_message_lines = [
+            "[green]Submission finished![/]",
+            "You can go back to job/session selection or quit BugIt.",
+        ]
 
         all_upload_ok = all(
             w.state == WorkerState.SUCCESS for w in self.upload_workers.values()
         )
-        if is_prod() and all_upload_ok:
+        if is_prod() and all_upload_ok and finalize_ok:
             # only cleanup if everything was uploaded
+            finish_message_lines.insert(
+                1,
+                f"URL: [$primary]{self.submitter.bug_url}[/]",
+            )
             shutil.rmtree(self.attachment_dir, ignore_errors=True)
 
-        finish_message_lines = [
-            "[green]Submission finished![/]",
-            f"URL: [$primary]{self.submitter.bug_url}[/]",
-            "You can go back to job/session selection or quit BugIt.",
-        ]
-
-        if not all_upload_ok and self.attachment_dir.exists():
+        if not (all_upload_ok and finalize_ok) and self.attachment_dir.exists():
             if is_snap():
                 attachment_dir = (
                     "/tmp/snap-private-tmp/snap.bugit-v2/tmp" / self.attachment_dir
